+++ pkg/controllers/componentUpgradeManager.go
@@ -11,25 +11,26 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
 
-// componentUpgradeManager is a central unit in the process of handling the upgrade process of a custom component resource.
-type componentUpgradeManager struct {
+// ComponentUpgradeManager is a central unit in the process of handling the upgrade process of a custom component resource.
+type ComponentUpgradeManager struct {
 	componentClient componentInterface
 	helmClient      helmClient
+	healthManager   healthManager
 	recorder        record.EventRecorder
 }
 
-// NewComponentUpgradeManager creates a new instance of componentUpgradeManager.
-func NewComponentUpgradeManager(componentClient componentInterface, helmClient helmClient, recorder record.EventRecorder) *componentUpgradeManager {
-	return &componentUpgradeManager{
+// NewComponentUpgradeManager creates a new instance of ComponentUpgradeManager.
+func NewComponentUpgradeManager(componentClient componentInterface, helmClient helmClient, healthManager healthManager, recorder record.EventRecorder) *ComponentUpgradeManager {
+	return &ComponentUpgradeManager{
 		componentClient: componentClient,
 		helmClient:      helmClient,
+		healthManager:   healthManager,
 		recorder:        recorder,
 	}
 }
 
 // Upgrade upgrades a given component resource.
-// nolint: contextcheck // uses a new non-inherited context to finish running helm-processes on SIGTERM
-func (cum *componentUpgradeManager) Upgrade(ctx context.Context, component *k8sv1.Component) error {
+func (cum *ComponentUpgradeManager) Upgrade(ctx context.Context, component *k8sv1.Component) error {
 	logger := log.FromContext(ctx)
 
 	err := cum.helmClient.SatisfiesDependencies(ctx, component.GetHelmChartSpec())
@@ -46,7 +47,8 @@ func (cum *componentUpgradeManager) Upgrade(ctx context.Context, component *k8sv
 	logger.Info("Upgrade helm chart...")
 
 	// create a new context that does not get canceled immediately on SIGTERM
-	helmCtx := context.Background()
+	// this allows self-upgrades
+	helmCtx := context.WithoutCancel(ctx)
 
 	if err := cum.helmClient.InstallOrUpgrade(helmCtx, component.GetHelmChartSpec()); err != nil {
 		return &genericRequeueableError{errMsg: fmt.Sprintf("failed to upgrade chart for component %s", component.Spec.Name), err: err}
@@ -55,6 +57,11 @@ func (cum *componentUpgradeManager) Upgrade(ctx context.Context, component *k8sv
 	component, err = cum.componentClient.UpdateStatusInstalled(helmCtx, component)
 	if err != nil {
 		return &genericRequeueableError{errMsg: fmt.Sprintf("failed to update status-installed for component %s", component.Spec.Name), err: err}
+	}
+
+	err = cum.healthManager.UpdateComponentHealth(helmCtx, component.Spec.Name, component.Namespace)
+	if err != nil {
+		return fmt.Errorf("failed to update health status for component %q: %w", component.Spec.Name, err)
 	}
 
 	logger.Info(fmt.Sprintf("Upgraded component %s.", component.Spec.Name))