+++ pkg/controllers/component_manager.go
@@ -9,39 +9,39 @@ import (
 	"k8s.io/client-go/tools/record"
 )
 
-// componentManager is a central unit in the process of handling component custom resources.
-// The componentManager creates, updates and deletes components.
-type componentManager struct {
+// DefaultComponentManager is a central unit in the process of handling component custom resources.
+// The DefaultComponentManager creates, updates and deletes components.
+type DefaultComponentManager struct {
 	installManager installManager
 	deleteManager  deleteManager
 	upgradeManager upgradeManager
 	recorder       eventRecorder
 }
 
-// NewComponentManager creates a new instance of componentManager.
-func NewComponentManager(clientset componentInterface, helmClient helmClient, recorder record.EventRecorder) *componentManager {
-	return &componentManager{
-		installManager: NewComponentInstallManager(clientset, helmClient, recorder),
+// NewComponentManager creates a new instance of DefaultComponentManager.
+func NewComponentManager(clientset componentInterface, helmClient helmClient, healthManager healthManager, recorder record.EventRecorder) *DefaultComponentManager {
+	return &DefaultComponentManager{
+		installManager: NewComponentInstallManager(clientset, helmClient, healthManager, recorder),
 		deleteManager:  NewComponentDeleteManager(clientset, helmClient),
-		upgradeManager: NewComponentUpgradeManager(clientset, helmClient, recorder),
+		upgradeManager: NewComponentUpgradeManager(clientset, helmClient, healthManager, recorder),
 		recorder:       recorder,
 	}
 }
 
 // Install installs the given component resource.
-func (m *componentManager) Install(ctx context.Context, component *k8sv1.Component) error {
+func (m *DefaultComponentManager) Install(ctx context.Context, component *k8sv1.Component) error {
 	m.recorder.Event(component, corev1.EventTypeNormal, InstallEventReason, "Starting installation...")
 	return m.installManager.Install(ctx, component)
 }
 
 // Delete deletes the given component resource.
-func (m *componentManager) Delete(ctx context.Context, component *k8sv1.Component) error {
+func (m *DefaultComponentManager) Delete(ctx context.Context, component *k8sv1.Component) error {
 	m.recorder.Event(component, corev1.EventTypeNormal, DeinstallationEventReason, "Starting deinstallation...")
 	return m.deleteManager.Delete(ctx, component)
 }
 
 // Upgrade upgrades the given component resource.
-func (m *componentManager) Upgrade(ctx context.Context, component *k8sv1.Component) error {
+func (m *DefaultComponentManager) Upgrade(ctx context.Context, component *k8sv1.Component) error {
 	m.recorder.Event(component, corev1.EventTypeNormal, UpgradeEventReason, "Starting upgrade...")
 	return m.upgradeManager.Upgrade(ctx, component)
 }