+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - [#51] Requeue components on errors during install, update and delete.
+### Added
+- [#48] Add labels with component name and version to all resources of a component
+- [#48] Track health on component CR
 
 ## [v0.7.0] - 2023-12-08
 ### Added